tool: rename sysxml to sdf
We refer to SDF (System Description Format) throughout the code
and documentation, this filename of the SDF parsing code should
reflect what it is for.
